# Conflicts:
#	pychron/canvas/canvas2D/laser_tray_canvas.py
#	pychron/lasers/stage_managers/stage_manager.py

+++ pychron/core/helpers/logger_setup.py
@@ -113,7 +113,7 @@ def set_exception_handler(func=None):
 #         logger.addHandler(h)
 
 
-def logging_setup(name, use_archiver=True, **kw):
+def logging_setup(name, use_archiver=True, use_file=True, **kw):
     """
     """
     # set up deprecation warnings
@@ -136,29 +136,35 @@ def logging_setup(name, use_archiver=True, **kw):
                      root=bdir)
         a.clean()
 
-    # create a new logging file
-    logname = '{}.current.log'.format(name)
-    logpath = os.path.join(bdir, logname)
+    if use_file:
+        # create a new logging file
+        logname = '{}.current.log'.format(name)
+        logpath = os.path.join(bdir, logname)
 
-    if os.path.isfile(logpath):
-        backup_logpath, _cnt = unique_path2(bdir, name, delimiter='-', extension='.log', width=5)
+        if os.path.isfile(logpath):
+            backup_logpath, _cnt = unique_path2(bdir, name, delimiter='-', extension='.log', width=5)
 
-        shutil.copyfile(logpath, backup_logpath)
-        os.remove(logpath)
+            shutil.copyfile(logpath, backup_logpath)
+            os.remove(logpath)
 
-        ps = list_directory(bdir, filtername=logname, remove_extension=False)
-        for pi in ps:
-            _h, t = os.path.splitext(pi)
-            v = os.path.join(bdir, pi)
-            shutil.copyfile(v, '{}{}'.format(backup_logpath, t))
-            os.remove(v)
+            ps = list_directory(bdir, filtername=logname, remove_extension=False)
+            for pi in ps:
+                _h, t = os.path.splitext(pi)
+                v = os.path.join(bdir, pi)
+                shutil.copyfile(v, '{}{}'.format(backup_logpath, t))
+                os.remove(v)
 
     root = logging.getLogger()
     root.setLevel(gLEVEL)
     shandler = logging.StreamHandler()
-    rhandler = RotatingFileHandler(
-        logpath, maxBytes=1e7, backupCount=5)
-    for hi in (shandler, rhandler):
+
+    handlers = [shandler]
+    if use_file:
+        rhandler = RotatingFileHandler(
+            logpath, maxBytes=1e7, backupCount=5)
+        handlers.append(rhandler)
+
+    for hi in handlers:
         hi.setLevel(gLEVEL)
         hi.setFormatter(logging.Formatter(gFORMAT))
         root.addHandler(hi)
